pyproject.toml
[build-system]
requires = [
    'setuptools>=42.0.0',
    'wheel >= 0.36.0',
    'pylint~=2.15.5',           # MIT
    'pytest~=7.2.0',            # GPL-2.0-or-later
    'pytest-cov~=4.0.0',        # MIT
    'pyjwt~=2.6.0',             # MIT
    'python-keycloak~=2.6.0'    # MIT
]
build-backend = 'setuptools.build_meta'

[project]
name = "mrmat-hello-python"
description = "Small Python projects that don't warrant their own repository"
urls = { "Sources" = "https://github.com/MrMatOrg/mrmat-hello-python" }
keywords = ["experimental"]
readme = "README.md"
license = { text = "MIT" }
authors = [
    { "name" = "Mathieu Imfeld", "email" = "imfeldma+9jqerw@gmail.com" }
]
maintainers = [
    { "name" = "Mathieu Imfeld", "email" = "imfeldma+9jqerw@gmail.com" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT",
    "Programming Language :: Python :: 3.10"
]
requires-python = ">=3.10"
dependencies = [
    "requests~=2.28.1",                                 # Apache 2.0
    "rich~=12.6.0",                                     # MIT
    "flask~=2.2.2",                                     # BSD 3-Clause
    "Flask-SQLAlchemy~=3.0.2",                          # BSD 3-Clause
    "Flask-Migrate~=3.1.0",                             # MIT
    "flask-smorest~=0.40.0",                            # MIT
    "Flask-Marshmallow~=0.14.0",                        # MIT
    "marshmallow-sqlalchemy~=0.28.0",                   # MIT
    "Flask-OIDC~=1.4.0",                                # MIT
    "psycopg2-binary~=2.9.1",                           # LGPL with exceptions
    "requests_oauthlib~=1.3.0",                         # ISC
    "cqlengine~=0.21.0",                                # BSD
    "mongoengine~=0.24.2",                              # MIT
    "pyspark~=3.3.1",                                   # Apache 2.0
    "imagehash~=4.3.1",                                 # BSD 2-Clause
    "piexif~=1.1.3",                                    # MIT
    'importlib-metadata; python_version<"3.10"',
]
dynamic = ["version"]

[tool.setuptools.dynamic]
version = { attr = "ci.version "}

[tool.setuptools.packages.find]
where = ["src"]
include = ["mhpython*"]
namespaces = false

[project.scripts]
mrmat-boilerplate = "mhpython.boilerplate.mrmat_boilerplate:run"
mrmat-classes = "mhpython.classes.mrmat_classes:run"
mrmat-github = "mhpython.github.mrmat_github:run"
mrmat-oidc-interactive = "mhpython.oidc.mrmat_oidc_interactive:run"
mrmat-cass-import = "mhpython.hackathon_2015.mrmat_cass_import:run"
mrmat-tweet-analyze = "mhpython.hackathon_2015.mrmat_tweet_analyze:run"
mrmat-localized = "mhpython.localization.localized:run"
mrmat-env = "mhpython.py_env.mrmat_env:run"
mrmat-io = "mhpython.py_io.mrmat_io:run"
mrmat-packages = "mhpython.py_packages.mrmat_packages:run"
mrmat-spinner = "mhpython.py_spinner.mrmat_spinner:run"
mrmat-pg-to-mongo = "mhpython.orm.mrmat_pg_to_mongo:run"
